+++ src/count.js
@@ -1,0 +1,5 @@
+const find = require('./find')
+
+module.exports = function count (tree, selector) {
+  return find(tree, selector).length
+}