+++ src/each.js
@@ -1,0 +1,5 @@
+const find = require('./find')
+
+module.exports = function each (tree, selector, callback) {
+  return find(tree, selector).forEach(callback)
+}