+++ src/find.js
@@ -1,0 +1,5 @@
+const esquery = require('esquery')
+
+module.exports = function find (tree, selector) {
+  return esquery(tree, selector)
+}