+++ src/first.js
@@ -1,0 +1,5 @@
+const find = require('./find')
+
+module.exports = function first (tree, selector) {
+  return find(tree, selector)[0]
+}