+++ src/generate.js
@@ -1,0 +1,5 @@
+const astring = require('astring')
+
+module.exports = function generate (tree, options) {
+  return astring.generate(tree, options)
+}