+++ src/has.js
@@ -1,0 +1,5 @@
+const count = require('./count')
+
+module.exports = function has (tree, selector) {
+  return count(tree, selector) > 0
+}