+++ src/last.js
@@ -1,0 +1,6 @@
+const find = require('./find')
+
+module.exports = function last (tree, selector) {
+  var nodes = find(tree, selector)
+  return nodes[nodes.length - 1]
+}