+++ src/mark.js
@@ -1,0 +1,9 @@
+const walk = require('./walk')
+
+module.exports = function mark (tree) {
+  let cid = 1
+  walk(tree, node => {
+    node.cid = cid
+    cid += 1
+  })
+}