+++ src/parse.js
@@ -1,0 +1,5 @@
+const cherow = require('cherow')
+
+module.exports = function parse (source, options) {
+  return cherow.parseModule(source, options)
+}