+++ src/prepend.js
@@ -1,0 +1,8 @@
+module.exports = function prepend (tree, node) {
+  if (Array.isArray(tree)) {
+    tree.unshift(node)
+  } else if (Array.isArray(tree.body)) {
+    tree.body.unshift(node)
+  }
+  return tree
+}