+++ src/traverse.js
@@ -1,0 +1,5 @@
+const estraverse = require('estraverse')
+
+module.exports = function traverse (tree, options) {
+  return estraverse.traverse(tree, options)
+}