+++ src/walk.js
@@ -1,0 +1,5 @@
+const estraverse = require('estraverse')
+
+module.exports = function walk (tree, callback) {
+  return estraverse.traverse(tree, { enter: callback })
+}