+++ test/basic.js
@@ -2,11 +2,11 @@ const test = require('./helpers/test')
 const AbstractSyntaxTree = require('..')
 
 test('it works', assert => {
-  assert.ok(new AbstractSyntaxTree('var x = 0;'))
+  assert.truthy(new AbstractSyntaxTree('var x = 0;'))
 })
 
 test('it accepts abstract syntax tree', assert => {
-  assert.ok(new AbstractSyntaxTree([{
+  assert.truthy(new AbstractSyntaxTree({
     type: 'Program',
     body: [
       {
@@ -16,13 +16,13 @@ test('it accepts abstract syntax tree', assert => {
         body: { 'type': 'BlockStatement', 'body': [] }
       }
     ]
-  }]))
+  }))
 })
 
 test('it queries the syntax tree', assert => {
   var ast = new AbstractSyntaxTree('var y = 1;')
   var declarations = ast.find('VariableDeclaration')
-  assert.ok(declarations.length === 1)
+  assert.truthy(declarations.length === 1)
 })
 
 test('it iterates over found nodes', assert => {
@@ -30,28 +30,28 @@ test('it iterates over found nodes', assert => {
   ast.each('Literal', node => {
     node.value = 2
   })
-  assert.ok(ast.first('Literal').value === 2)
+  assert.truthy(ast.first('Literal').value === 2)
 })
 
 test('it checks if node is in the syntax tree', assert => {
   var ast = new AbstractSyntaxTree('var z = 2;')
-  assert.ok(ast.has('VariableDeclaration'))
+  assert.truthy(ast.has('VariableDeclaration'))
 })
 
 test('it counts nodes', assert => {
   var ast = new AbstractSyntaxTree('var z = 2; var x = 3;')
-  assert.ok(ast.count('VariableDeclaration') === 2)
+  assert.truthy(ast.count('VariableDeclaration') === 2)
 })
 
 test('it returns the source', assert => {
   var ast = new AbstractSyntaxTree('var a = 3;')
-  assert.ok(ast.toSource() === 'var a = 3;\n')
+  assert.truthy(ast.source === 'var a = 3;\n')
 })
 
 test('it removes nodes', assert => {
   var ast = new AbstractSyntaxTree('"use strict"; var b = 4;')
   ast.remove({ type: 'Literal', value: 'use strict' })
-  assert.ok(ast.toSource() === 'var b = 4;\n')
+  assert.truthy(ast.source === 'var b = 4;\n')
 })
 
 test('it removes function declarations', assert => {
@@ -64,7 +64,7 @@ test('it removes function declarations', assert => {
       name: 'hello'
     }
   })
-  assert.ok(ast.toSource() === 'var a = 1;\n')
+  assert.truthy(ast.source === 'var a = 1;\n')
 })
 
 test('it keeps variable declarations', assert => {
@@ -77,7 +77,7 @@ test('it keeps variable declarations', assert => {
       name: 'a'
     }
   })
-  assert.ok(ast.toSource() === 'var b = 2;\nfunction hello() {\n  return "world";\n}\n')
+  assert.truthy(ast.source === 'var b = 2;\nfunction hello() {\n  return "world";\n}\n')
 })
 
 test('it removes empty declarations', assert => {
@@ -90,64 +90,52 @@ test('it removes empty declarations', assert => {
       name: 'a'
     }
   })
-  assert.ok(ast.toSource() === 'function hello() {\n  return "world";\n}\n')
+  assert.truthy(ast.source === 'function hello() {\n  return "world";\n}\n')
 })
 
 test('it removes nodes if string is provided', assert => {
   var source = 'var a = 1; function hello () { return "world"; }'
   var ast = new AbstractSyntaxTree(source)
   ast.remove('VariableDeclaration')
-  assert.deepEqual(ast.toSource(), 'function hello() {\n  return "world";\n}\n')
+  assert.deepEqual(ast.source, 'function hello() {\n  return "world";\n}\n')
 })
 
 test('it removes nodes via complex selectors', assert => {
   var source = 'var a = 1; function hello () { var b = 2; return "world"; }'
   var ast = new AbstractSyntaxTree(source)
   ast.remove('BlockStatement > VariableDeclaration')
-  assert.deepEqual(ast.toSource(), `var a = 1;\nfunction hello() {\n  return "world";\n}\n`)
+  assert.deepEqual(ast.source, `var a = 1;\nfunction hello() {\n  return "world";\n}\n`)
 })
 
 test('it returns the first node', assert => {
   var ast = new AbstractSyntaxTree('var a = 1; var b = 2;')
   var declaration = ast.first('VariableDeclaration')
-  assert.ok(declaration.declarations[0].id.name === 'a')
+  assert.truthy(declaration.declarations[0].id.name === 'a')
 })
 
 test('it returns the last node', assert => {
   var ast = new AbstractSyntaxTree('var c = 3; var d = 4;')
   var declaration = ast.last('VariableDeclaration')
-  assert.ok(declaration.declarations[0].id.name === 'd')
+  assert.truthy(declaration.declarations[0].id.name === 'd')
 })
 
 test('it works with imports', assert => {
   var source = 'import foo from "bar";'
   var ast = new AbstractSyntaxTree(source)
-  assert.deepEqual(ast.toSource(), source + '\n')
+  assert.deepEqual(ast.source, source + '\n')
 })
 
 test('it should be possible to remove the first element only', assert => {
   var source = 'var a = 1; var b = 2;'
   var ast = new AbstractSyntaxTree(source)
   ast.remove({ type: 'VariableDeclaration' }, { first: true })
-  assert.ok(ast.toSource() === 'var b = 2;\n')
-})
-
-test('it should be possible to beautify the source', assert => {
-  var source = 'var a = 1; var b = 2;'
-  var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.toSource({ beautify: true }) === 'var a = 1;\nvar b = 2;\n')
-})
-
-test('it accepts parameters for the beautify method', assert => {
-  var source = 'var x = "y";'
-  var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.toSource({ beautify: { semi: false } }) === 'var x = "y"\n')
+  assert.truthy(ast.source === 'var b = 2;\n')
 })
 
 test('it supports double quotes by default', assert => {
   var source = `var a = 'hello';`
   var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.toSource() === `var a = "hello";\n`)
+  assert.truthy(ast.source === `var a = "hello";\n`)
 })
 
 test('it prepends a node to body', assert => {
@@ -160,7 +148,7 @@ test('it prepends a node to body', assert => {
       value: 'use strict'
     }
   })
-  assert.deepEqual(ast.toSource(), '"use strict";\nvar a = 1;\n')
+  assert.deepEqual(ast.source, '"use strict";\nvar a = 1;\n')
 })
 
 test('it appends a node to body', assert => {
@@ -173,13 +161,7 @@ test('it appends a node to body', assert => {
       value: 'use strict'
     }
   })
-  assert.deepEqual(ast.toSource(), 'var a = 1;\n"use strict";\n')
-})
-
-test('it compares nodes', assert => {
-  var source = 'var a = 1;'
-  var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.is({ type: 'ReturnStatement' }, { type: 'ReturnStatement' }))
+  assert.deepEqual(ast.source, 'var a = 1;\n"use strict";\n')
 })
 
 test('it walks through nodes', assert => {
@@ -191,7 +173,7 @@ test('it walks through nodes', assert => {
     }
     return node
   })
-  assert.deepEqual(ast.toSource(), 'let a = 1;\n')
+  assert.deepEqual(ast.source, 'let a = 1;\n')
 })
 
 test('it exposes a traverse method', assert => {
@@ -205,7 +187,7 @@ test('it exposes a traverse method', assert => {
       return node
     }
   })
-  assert.deepEqual(ast.toSource(), 'let a = 1;\n')
+  assert.deepEqual(ast.source, 'let a = 1;\n')
 })
 
 test('it replaces nodes', assert => {
@@ -219,37 +201,12 @@ test('it replaces nodes', assert => {
       return node
     }
   })
-  assert.deepEqual(ast.toSource(), 'let a = 1;\n')
+  assert.deepEqual(ast.source, 'let a = 1;\n')
 })
 
 test('it has a toString alias for toSource', assert => {
   var ast = new AbstractSyntaxTree('var y = 1;')
-  assert.ok(ast.toString() === 'var y = 1;\n')
-})
-
-test('it allows to override the beautify method', assert => {
-  var source = 'var a = 1;'
-  var ast = new AbstractSyntaxTree(source)
-  ast.beautify = function (source) {
-    return source + '\n'
-  }
-  assert.deepEqual(ast.toSource({ beautify: true }), 'var a = 1;\n\n')
-})
-
-test('it has a minify noop method by default', assert => {
-  var source = 'var hello = 1;'
-  var ast = new AbstractSyntaxTree(source)
-  assert.deepEqual(ast.toSource({ minify: true }), 'var hello = 1;\n')
-})
-
-test('it allows to override the minify method', assert => {
-  var source = 'var hello = 1;'
-  var ast = new AbstractSyntaxTree(source)
-  ast.minify = function (ast) {
-    ast.body[0].declarations[0].id.name = 'a'
-    return ast
-  }
-  assert.deepEqual(ast.toSource({ minify: true }), 'var a = 1;\n')
+  assert.truthy(ast.source === 'var y = 1;\n')
 })
 
 test('it wraps the code', assert => {
@@ -279,45 +236,33 @@ test('it wraps the code', assert => {
       }
     ]
   })
-  assert.deepEqual(ast.toSource().replace(/\s/g, ''), '(function(){vara=1;})();')
+  assert.deepEqual(ast.source.replace(/\s/g, ''), '(function(){vara=1;})();')
 })
 
 test('it unwraps the code in case of iife', assert => {
   var source = '(function () { console.log(1); }());'
   var ast = new AbstractSyntaxTree(source)
   ast.unwrap()
-  assert.ok(ast.toSource() === 'console.log(1);\n')
+  assert.truthy(ast.source === 'console.log(1);\n')
 })
 
 test('it unwraps code in case of amd', assert => {
   var source = 'define(function () { console.log(1); });'
   var ast = new AbstractSyntaxTree(source)
   ast.unwrap()
-  assert.ok(ast.toSource() === 'console.log(1);\n')
+  assert.truthy(ast.source === 'console.log(1);\n')
+})
+
+test('it returns the type', assert => {
+  var source = 'var a = 1;'
+  var ast = new AbstractSyntaxTree(source)
+  assert.deepEqual(ast.type, 'Program')
 })
 
 test('it returns the body', assert => {
   var source = 'var a = 1;'
   var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.body() === ast.ast.body)
-})
-
-test('it generates ast from templates', assert => {
-  var source = 'console.log(1);'
-  var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.template('"use strict";')[0].type === 'ExpressionStatement')
-})
-
-test('it generates ast from templates with parameters', assert => {
-  var source = 'console.log(1);'
-  var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.template('var x = <%= value %>;', { value: { type: 'Literal', value: 1 } })[0].declarations[0].init.value === 1)
-})
-
-test('it generates ast from literals', assert => {
-  var source = 'console.log(1);'
-  var ast = new AbstractSyntaxTree(source)
-  assert.ok(ast.template(1).type === 'Literal')
+  assert.truthy(ast.body)
 })
 
 test('it generates sourcemaps', assert => {
@@ -325,18 +270,16 @@ test('it generates sourcemaps', assert => {
   ast.each('Literal', node => {
     node.value = 2
   })
-  assert.ok(ast.first('Literal').value === 2)
-  const { source, map } = ast.toSource({ sourceMap: true })
-  assert.ok(source)
-  assert.deepEqual(map, '{"version":3,"sources":["UNKNOWN"],"names":["y"],"mappings":"IAAIA,IAAI","file":"UNKNOWN"}')
+  assert.truthy(ast.first('Literal').value === 2)
+  assert.deepEqual(ast.map, '{"version":3,"sources":["UNKNOWN"],"names":["y"],"mappings":"IAAIA,IAAI","file":"UNKNOWN"}')
 })
 
 test('it lets you mark nodes', assert => {
   var ast = new AbstractSyntaxTree('var a = 1;')
   ast.mark()
-  assert.ok(ast.first('Program').cid === 1)
-  assert.ok(ast.first('VariableDeclaration').cid === 2)
-  assert.ok(ast.first('VariableDeclarator').cid === 3)
-  assert.ok(ast.first('Identifier').cid === 4)
-  assert.ok(ast.first('Literal').cid === 5)
+  assert.truthy(ast.first('Program').cid === 1)
+  assert.truthy(ast.first('VariableDeclaration').cid === 2)
+  assert.truthy(ast.first('VariableDeclarator').cid === 3)
+  assert.truthy(ast.first('Identifier').cid === 4)
+  assert.truthy(ast.first('Literal').cid === 5)
 })