+++ test/helpers/test.js
@@ -1,14 +1,3 @@
-const assert = require('assert')
-
-async function test (description, callback) {
-  try {
-    await callback(assert)
-  } catch (exception) {
-    console.error(exception)
-    console.log(`Test failed: ${description}`)
-  }
-}
-
-test.skip = function () {}
+const test = require('ava')
 
 module.exports = test