+++ test/jsx.js
@@ -2,5 +2,5 @@ const test = require('./helpers/test')
 const AbstractSyntaxTree = require('..')
 
 test('it works for jsx', assert => {
-  assert.ok(new AbstractSyntaxTree('var x = (<div>foo</div>);', { jsx: true }))
+  assert.truthy(new AbstractSyntaxTree('var x = (<div>foo</div>);', { jsx: true }))
 })